+++ src/shared/constants/app.ts
@@ -11,3 +11,5 @@ export const AFTER_LOGIN_REDIRECT_PATH = PATH.DASHBOARD;
 /** config */
 export const FILTER_STALE_TIME = Infinity;
 export const RETRY_COUNT = 0;
+export const ACCESS_TOKEN_EXPIRE = 60 * 60 * 1000; // 1시간
+export const REFRESH_TOKEN_EXPIRE = 7 * 24 * 60 * 60 * 1000; // 7일
